requirements.txt
# Core dependencies
elevenlabs>=0.2.27
openai>=1.12.0
gradio>=4.19.2
python-dotenv>=1.0.1
requests>=2.31.0
google-generativeai>=0.5.0
markitdown>=0.1.1
tiktoken>=0.5.1

# Audio processing
pydub>=0.25.1
ffmpeg-python>=0.2.0
whisper>=1.1.10
numpy>=1.19.3
torch>=2.0.0

# Networking and utilities
urllib3>=2.0.0
typing-extensions>=4.7.0

# Web framework
fastapi>=0.109.2
python-multipart>=0.0.9
streamlit==1.32.0

# Image processing
pillow>=9.0.0
pillow-heif>=0.13.0

# Utilities
tqdm>=4.66.2 